refactor: public and private routes
