+++ app/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import HomePageComponents from './home/page-components'
+import HomePageComponents from './_home/page-components'
 
 export default async function HomePage() {
